chore: LAST ONE, renames to make genkit service explicit
